LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/shared-workers.https-expected.txt

Verify the Cross-Origin Embedder Policy for Shared Workers by performing a cross-domain "fetch" request for a resource that does not specify a COEP. Only Shared Workers with the default COEP should be able to successfully perform this operation.


PASS default policy (derived from response)
PASS "require-corp" (derived from response)
PASS default policy (derived from owner set due to use of local scheme - blob URL)
FAIL "require-corp" (derived from owner set due to use of local scheme - blob URL) assert_equals: expected "failure" but got "success"
PASS default policy (derived from owner set due to use of local scheme - data URL)
FAIL "require-corp" (derived from owner set due to use of local scheme - data URL) assert_equals: expected "failure" but got "success"
